Fixed bullet firing logic and added item
This change fixes the bug where the bullet's y position was
being set to the player's y position because it was a child of
the player. This change also adds an item that gets spawned
sometimes when an enemy is killed. However, there is a bug where
the position of the item is incorrect