Added xp and hp bar
Added xp and hp bar to a new MainUi scene. This scene will eventually
hold all of the main ui on screen.
This change also removes player collision with enemies. Instead, we
want the player to collide with the enemy hitbox and take damage. The
enemies will still be colliding with each other.
Also removed health component from the player since it is easier to
track player health globally.